ocr/requirements.txt
numpy==1.24.4
Pillow>=10.1.0,<11.0.0
marker-pdf>=0.2.0
pdf2docx==0.5.6
PyMuPDF==1.20.2
python-docx==1.1.0
matplotlib==3.7.2
underthesea
pyvi
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66